requirements.txt
pandas
streamlit
PyPDF2
python-docx
python-pptx
pillow
langchain
python-dotenv
